refactor(backend): extract OAuth refresh classification to a service module
The OAuth refresh-outcome classification and error-sanitization helpers
(classifyRefreshResponse, classifyThrownRefreshError,
refreshFailureToServerFields, sanitizeOAuthErrorCode,
sanitizeOAuthErrorDescription, and the OAuthRefreshOutcome type) were pure
functions living in the oauth route file, with their unit tests in the
route-level test file. Move them to src/services/oauth-refresh-classification.ts
with a colocated unit test, so route wiring and pure business logic are
separated and the tests run in the mock-free vitest project. No behavior
change; refreshOAuthToken (which does the actual I/O) stays in the route and
consumes the service.
